backend/pyproject.toml
[project]
name = "ixora-meeting-booking"
version = "0.1.0"
description = "AI agent for booking meetings with Ixora Solution's CEO via Microsoft Bookings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.0.20",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.22",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pypdf>=4.0.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "playwright>=1.40.0",
]
